Work in progress

+++ wipesrc/win95/origsrc/fileops.h
@@ -1,7 +1,7 @@
 typedef struct _TIMlist
 {
 	void *memBase;
-	long	memOffset[400];
+	intptr_t	memOffset[400];
 } TIMlist;
 
 extern TIMlist *LoadCompressedTextureSequence( char *texName );
