+++ wipesrc/wipeout98/fileops.h
@@ -1,7 +1,9 @@
+#include <stdint.h>
+
 typedef struct _TIMlist
 {
 	void *memBase;
-	long	memOffset[400];
+	uintptr_t	memOffset[400];
 } TIMlist;
 
 extern TIMlist *LoadCompressedTextureSequence( char *texName );
